feat: new balancer vault contract and tests
