+++ src/index.ts
@@ -14,7 +14,6 @@ import PrimaryIssueManager from './contract/amm/primary';
 import SecondaryIssueManager from './contract/amm/secondary';
 import MarginIssueManager from './contract/amm/margin';
 import Security from './contract/security';
-import BalancerVault from './contract/BalancerVault';
 import SecuritiesFactory from './contract/securitiesfactory';
 import Distribution from './contract/distribution';
 import Client from './contract/client';
@@ -37,7 +36,6 @@ export {
     SecondaryIssueManager,
     MarginIssueManager,
     Security,
-    BalancerVault,
     SecuritiesFactory,
     Distribution,
     Client,
